requirements.txt
langchain==0.1.0
langgraph==0.0.10
openai==1.3.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
facebook-sdk==3.1.0
pydantic==2.5.2
requests==2.31.0